requirements.txt
aiofiles==0.6.0
async-exit-stack==1.0.1
async-generator==1.10
click==7.1.2
dataclasses==0.8
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
idna==3.1
importlib-metadata==4.0.1
Jinja2==3.0.0
lockfile==0.12.2
MarkupSafe==2.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
zipp==3.4.1
